+++ src/lib/builtins/mod.rs
@@ -1,5 +1,7 @@
 /// The global `Array` object
 pub mod array;
+/// the global `Symbol` Object
+pub mod symbol;
 // The global `Boolean` object
 pub mod boolean;
 /// The global `console` object
@@ -12,13 +14,15 @@ pub mod function;
 pub mod json;
 /// The global `Math` object
 pub mod math;
+/// The global `Number` object
+pub mod number;
 /// The global `Object` object
 pub mod object;
+/// Property, used by `Object`
+pub mod property;
 /// The global 'RegExp' object
 pub mod regexp;
 /// The global `String` object
 pub mod string;
 /// Javascript values, utility methods and conversion between Javascript values and Rust values
 pub mod value;
-// Property, used by `Object`
-pub mod property;