+++ src/lib/environment/object_environment_record.rs
@@ -72,7 +72,7 @@ impl EnvironmentRecordTrait for ObjectEnvironmentRecord {
 
     fn get_binding_value(&self, name: &str, strict: bool) -> Value {
         if self.bindings.has_field(name) {
-            self.bindings.get_field(name)
+            self.bindings.get_field_slice(name)
         } else {
             if strict {
                 // TODO: throw error here