+++ tests/js/test.js
@@ -1,6 +1,1 @@
-///
-// Use this file to test your changes to boa.
-///
-
-let a = Boolean(0);
-typeof a;
+// Test your JS here